Form Validation part 2 - task 2 completed

+++ SpringCMS/src/main/java/pl/simo333/repository/CategoryDao.java
@@ -1,6 +1,7 @@
 package pl.simo333.repository;
 
 import org.hibernate.Hibernate;
+import org.hibernate.validator.internal.engine.ConstraintViolationImpl;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Propagation;
 import pl.simo333.domain.Author;
@@ -9,7 +10,9 @@ import pl.simo333.domain.Category;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
+import javax.validation.ConstraintViolationException;
 import java.util.List;
+import java.util.Set;
 
 @Repository
 @Transactional
@@ -34,14 +37,12 @@ public class CategoryDao {
         entityManager.merge(category);
     }
 
-    public void delete(Category category) {
-        Category attached = entityManager
-                .createQuery("SELECT c FROM Category c JOIN FETCH c.articles WHERE c.id = ?1", Category.class)
-                .setParameter(1, category.getId())
-                .getSingleResult();
-        attached.getArticles().forEach(article -> article.removeCategory(category));
-
-        entityManager.remove(entityManager.contains(attached) ?
-                attached : entityManager.merge(attached));
+    public void delete(Long id) {
+        Category attached = findById(id);
+        if (attached.getArticles().isEmpty()) {
+            entityManager.remove(entityManager.contains(attached) ?
+                    attached : entityManager.merge(attached));
+        }
+        throw new IllegalStateException("Cannot remove due to articles attached.");
     }
 }
